Amend specs depending on SR4R feature flag
