+++ spec/requests/provider_interface/rejections_controller_spec.rb
@@ -24,8 +24,6 @@ RSpec.describe ProviderInterface::RejectionsController, type: :request do
           last_name: provider_user.last_name,
         ),
       )
-
-    FeatureFlag.activate(:structured_reasons_for_rejection_redesign)
   end
 
   describe 'GET edit' do
@@ -43,18 +41,6 @@ RSpec.describe ProviderInterface::RejectionsController, type: :request do
       let(:status) { 'rejected' }
 
       it 'responds with 404' do
-        get new_provider_interface_rejection_path(application_choice)
-
-        expect(response.status).to eq(404)
-      end
-    end
-
-    context 'if feature flag is not active' do
-      let(:status) { 'awaiting_provider_decision' }
-
-      it 'responds with 404' do
-        FeatureFlag.deactivate(:structured_reasons_for_rejection_redesign)
-
         get new_provider_interface_rejection_path(application_choice)
 
         expect(response.status).to eq(404)
